WiFi Module: Add AccessPoint mode
Sometimes, the station mode does seems to properly work.
UDP broadcasts are not transmitted and we get weird connection
issues.

It is currently unknown why, but configuring the ESP to run in AP
mode instead. Then, the ESP creates a WiFi network on its own that
a remote client can connect to. This seems to be stable.